Can now run an effects chain with e.g. upsample, filter, downsample, dither. Cleaned up speed hack so as not to affect resampling effects.

+++ src/echo.c
@@ -141,7 +141,7 @@ static int sox_echo_start(sox_effect_t * effp)
                 if ( echo->samples[i] > (sox_ssize_t)DELAY_BUFSIZ )
                 {
                         sox_fail("echo: delay must be less than %g seconds!",
-                                DELAY_BUFSIZ / (float) effp->ininfo.rate );
+                                DELAY_BUFSIZ / effp->ininfo.rate );
                         return (SOX_EOF);
                 }
                 if ( echo->decay[i] < 0.0 )
