+++ src/ladspa.c
@@ -198,8 +198,8 @@ static int sox_ladspa_start(sox_effect_t * effp)
   unsigned long i;
 
   /* Instantiate the plugin */
-  sox_debug("rate for plugin is %d", effp->ininfo.rate);
-  l_st->handle = l_st->desc->instantiate(l_st->desc, effp->ininfo.rate);
+  sox_debug("rate for plugin is %g", effp->ininfo.rate);
+  l_st->handle = l_st->desc->instantiate(l_st->desc, (unsigned long)effp->ininfo.rate);
   if (l_st->handle == NULL) {
     sox_fail("could not instantiate plugin");
     return SOX_EOF;
